projeto2/requirements.txt
Flask>=2.3.2
Flask-SQLAlchemy>=3.1.1
Flask-Cors>=3.0.10
Werkzeug>=2.3.8
flask-swagger-ui>=4.11.1
flask-login>=0.6.3
pytest>=7.4.3
python-dotenv>=1.0.1
pymysql
cryptography
bcrypt
# yse `pip show <package-name>` to show current package details and version
# use `pip freeze > requirements.txt` to add current packages do requirements.txt
